# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/employees/roles": {
      "get": {
        "operationId": "EmployeeProfilesController_getDepartments_v1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeRolesDto"
                }
              }
            }
          },
          "404": {
            "description": "No roles found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Roles: organization] Get students academic information",
        "tags": [
          "Employees"
        ]
      }
    },
...
    "/api/v1/employees/bulk-create/file": {
      "post": {
        "operationId": "EmployeeProfilesController_uploadEmployeesProfiles_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "operationId": "EmployeeProfilesController_getStudentDetail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Employee not found"
          },
          "404": {
            "description": "Employee not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Roles: organization] Get employee detail by ID",
        "tags": [
          "Employees"
        ]
      }
    },
...
    "/api/v1/data-wipe/preview/{organizationId}": {
      "get": {
        "description": "Shows what screening and counseling data will be deleted for users in the specified organization, without actually deleting anything.",
        "operationId": "DataWipeController_previewWipeData_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preview data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Wipe preview generated successfully",
                    "data": {
                      "organizationId": "ad6fb6dd-fa6f-46c5-ba76-9a2e5c6bcbcc",
                      "totalUsers": 15,
                      "usersByRole": {
                        "employees": 12,
                        "students": 3,
                        "others": 0
                      },
                      "screeningsCount": 45,
                      "counselingsCount": 8,
                      "users": [
                        {
                          "id": "user-123",
                          "fullName": "John Doe",
                          "email": "john@company.com",
                          "role": "employee",
                          "screeningsCount": 3,
                          "counselingsCount": 1
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview data that will be wiped for an organization",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/execute/{organizationId}": {
      "delete": {
        "description": "PERMANENTLY DELETES all screening and counseling data for users in the specified organization. This action cannot be undone!",
        "operationId": "DataWipeController_executeWipe_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data wipe executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Successfully wiped screening data for organization ad6fb6dd-fa6f-46c5-ba76-9a2e5c6bcbcc",
                    "data": {
                      "organizationId": "ad6fb6dd-fa6f-46c5-ba76-9a2e5c6bcbcc",
                      "deletedData": {
                        "screenings": 45,
                        "counselings": 8
                      },
                      "affectedUsers": 15,
                      "timestamp": "2024-01-15T10:30:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🚨 DANGER: Execute data wipe for an organization",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/stats/{organizationId}": {
      "get": {
        "description": "Retrieves current statistics about screening and counseling data for users in the specified organization.",
        "operationId": "DataWipeController_getOrganizationStats_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization statistics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get screening statistics for an organization",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/preview/current-org": {
      "get": {
        "description": "Shows what screening and counseling data will be deleted for users in the current user's organization.",
        "operationId": "DataWipeController_previewCurrentOrgWipeData_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview data that will be wiped for current user's organization",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/stats/current-org": {
      "get": {
        "description": "Retrieves current statistics about screening and counseling data for the current user's organization.",
        "operationId": "DataWipeController_getCurrentOrgStats_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get screening statistics for current user's organization",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/preview/target-organization": {
      "get": {
        "description": "Convenience endpoint to preview wipe data for the organization ID specified in the task.",
        "operationId": "DataWipeController_previewTargetOrgData_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview data for the target organization (ad6fb6dd-fa6f-46c5-ba76-9a2e5c6bcbcc)",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/execute/target-organization": {
      "delete": {
        "description": "PERMANENTLY DELETES all screening and counseling data for the organization ID specified in the task. This action cannot be undone!",
        "operationId": "DataWipeController_executeTargetOrgWipe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🚨 DANGER: Execute wipe for target organization (ad6fb6dd-fa6f-46c5-ba76-9a2e5c6bcbcc)",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
    "/api/v1/data-wipe/stats/target-organization": {
      "get": {
        "description": "Get current statistics for the organization ID specified in the task.",
        "operationId": "DataWipeController_getTargetOrgStats_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get statistics for target organization (ad6fb6dd-fa6f-46c5-ba76-9a2e5c6bcbcc)",
        "tags": [
          "Data Wipe Operations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ruang Diri API",
    "description": "API documentation for Ruang Diri application, all the services belongs to PT. Wong Makmur Sejati",
    "version": "1.0",
    "contact": {
      "name": "Support Team",
      "url": "https://ruangdiri.id",
      "email": "support@ruangdiri.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://ruangdiri.id/terms"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token here"
      }
    },
    "schemas": {
...
      "EmployeeRolesDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Information about roles and departments of employees",
            "example": {
              "departments": [
                "Human Resources",
                "Engineering",
                "Marketing",
                "Sales",
                "Finance"
              ],
              "position": [
                "Manager",
                "Software Engineer",
                "Marketing Specialist",
                "Sales Executive",
                "Accountant"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
      "EmployeeDetailDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Detailed information about an employee",
            "example": {
              "id": "7665717b-031b-4508-817c-92aa0423c09b",
              "email": "employee5@example.com",
              "profilePicture": "https://b1fbf745df68d22238d898899f1e0a87.serveo.net/uploads/organizations/profilePicture-1747884491936-759900585.jpg",
              "fullName": "Lewis Hamilton",
              "role": "employee",
              "employeeProfile": {
                "employeeId": "",
                "department": "Sales",
                "position": "Engineer",
                "gender": "male",
                "age": 26,
                "yearsOfService": 1,
                "screeningStatus": "stable",
                "counselingStatus": false,
                "guardianName": "Parent 5",
                "guardianContact": "081231001",
                "birthDate": "2002-01-01T00:00:00.000Z",
                "birthPlace": "City Center"
              },
              "mentalHealthHistories": [
                {
                  "id": "f69db26f-befc-47a4-8e23-61352e1374b8",
                  "type": "screening",
                  "date": "2025-03-09T04:30:02.459Z",
                  "status": "stable",
                  "notes": null
                },
                {
                  "id": "7a796386-576f-438f-a3a1-d147ea95ace0",
                  "type": "screening",
                  "date": "2025-03-05T07:30:59.467Z",
                  "status": "stable",
                  "notes": null
                },
                {
                  "id": "2ceca994-4b97-41dc-ae70-c85da43e123d",
                  "type": "screening",
                  "date": "2025-05-11T07:30:59.467Z",
                  "status": "monitored",
                  "notes": null
                },
                {
                  "id": "2f43dbcd-198d-4211-b4b3-962b114b6f97",
                  "type": "screening",
                  "date": "2025-04-04T07:30:59.467Z",
                  "status": "monitored",
                  "notes": null
                }
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}